pyproject.toml
[project]
name = "media-re-encode"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "Pillow>=11.3.0",
    "ffmpeg-python>=0.2.0",
]
